+++ crypto/__init__.py
@@ -1,10 +1,12 @@
-import fire
-from primitives.sha import sha3_224, sha3_256, sha3_512
-from primitives.authenticated_encryption import generate_key, AESGCM_encrypt, AESGCM_decrypt
-from primitives.signatures import generates_key_pair, sign, verify
-from primitives.homomorphic import he_encrypt, he_sum, he_decrypt, he_keys
+"""Init file for the crypto module."""
 
-class Crypto(object):
+import fire
+from crypto.sha import sha3_224, sha3_256, sha3_512
+from crypto.authenticated_encryption import generate_key, AESGCM_encrypt, AESGCM_decrypt
+from crypto.signatures import generates_key_pair, sign, verify
+from crypto.homomorphic import he_encrypt, he_sum, he_decrypt, he_generate_keys
+
+class Crypto:
     """Exposes cryptographic methods"""
 
     sha3_224           = sha3_224
@@ -18,12 +20,8 @@ class Crypto(object):
     generates_key_pair = generates_key_pair
     sign               = sign
     verify             = verify
-    
-    he_keys            = he_keys
+
+    he_generate_keys   = he_generate_keys
     he_encrypt         = he_encrypt
     he_sum             = he_sum
     he_decrypt         = he_decrypt
-
-
-if __name__ == '__main__':
-    fire.Fire(Crypto)
