+++ crypto/__main__.py
@@ -1,0 +1,7 @@
+"""main file for the crypto module."""
+
+import fire
+from . import Crypto
+
+if __name__ == "__main__":
+    fire.Fire(Crypto)
