+++ crypto/authenticated_encryption.py
