--- kenc.json
@@ -1,5 +1,0 @@
-{
-    "a": 1849370625423100704013131008240115598980772353736573775177651256479530951167175019864274856854952306521244044260023720569625400901225981071272240954424990306524745256182801960285217882467357127306060832580115924316613797291631185507981559664248385128573440542343744556356950614701962320287000604373512231684819953166981777187981630404657834609254935611034965194338899468619560345246316178208660824152191179935192349516371658875214935460346955613476541237822593086301319156125437844087802940357077196361818785131617917537250747350759879939897385457890775009143397094486198286933575460130033037332778307848565109116904,
-    "b": 5168818839331805460957189687064673469437063497435895506799129460529797339011246538768483232711613806083660215630331056976072903294279182044876947937612171601855426197168149018907721865604137953423166288056698851110964990280145658934300847977580630175557445830698344492128304012138871387542991348909754443864194821534766486629568664226281329973006094780278150171523627679461291369190236383001412248583213216304327151641837883181302855909248422983922850049565680092100306468425872278720796079097205693420243296408774716619500397725454792487540610206094086692781073084561127445208221964967866010536333152702812076162343,
-    "sum": 4665834883271005990387953805048252786255046719976718585563184645867334814580761023857431919861964784739502874304471204109325178218580326799046031101199179811329218155732291877042221541919495462353099893106347796594400647184074327281237739349943446593104233427419267307750289879700236126871313704402569112888798479452905804882648283784625245746695328305031123572864180566277181994839810687453779861940741263909313678684344140357398194921944189604936354771898898620658037780399161951644576236910079355073876901564330669846756455956738322702724824873298786664354970753578729217188002180293364669673184836615159610497179
-}
